add checksums in the plugin instead of requiring that in the build script
